python/pyproject.toml
[project]
name = "Ranker-Sorter"
version = "1.0.0"
description = ""
authors = [{ name = "Samuel Wu", email = "twopizza9621536@gmail.com" }]
dependencies = [
    "dearpygui>=1.8.0",
    "Python-Elo-System @ git+https://github.com/TwoPizza9621536/Python-Elo-System/#subdirectory=python",
]
requires-python = ">=3.10"
license = { text = "MIT" }

[tool.pdm]

[tool.pdm.dev-dependencies]
lint = ["black", "pylint", "flake8"]

[tool.pylint.main]
py-version = "3.10"
recursive = true

[tool.pylint.format]
max-line-length = 79

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
